Changes: - Changed Price to use Double instead of Int due to the need for floating point calculations - Changed CalculationUnit to Double due to the same reasons Finished: - Gas class Added: - Requests, a simple request handler for (as of now) handling get requests and converting strings to jsonObjects

+++ src/jvmMain/kotlin/CalculationUnit.kt
@@ -1,5 +1,5 @@
 interface CalculationUnit {
-    fun calculate(amount: Int) :Price
+    fun calculate(amount: Double) :Price
 
     override fun toString(): String
 }
